+++ src/app/other-user-skills/other-user-skills.component.scss
@@ -1,0 +1,8 @@
+@import "variables";
+
+.other-user-skills-error {
+  color: $myskills-error-message-text-color;
+  font-size: 14px;
+  margin-bottom: 16px;
+  white-space: pre-wrap;
+}
