+++ src/app/skill-card/skill-view.ts
@@ -1,0 +1,4 @@
+export interface SkillView {
+  id: string;
+  name: string;
+}
