+++ src/app/skill-card/user-skill-view.ts
@@ -1,0 +1,8 @@
+import { SkillView } from './skill-view';
+
+export interface UserSkillView {
+  skill: SkillView;
+  currentLevel: number;
+  desiredLevel: number;
+  priority: number;
+}
